Unify Klarna scenarios
