--- projects/magento/tests/KlarnaPayLaterPayment.spec.js
@@ -1,34 +1,0 @@
-import { test } from "@playwright/test";
-import KlarnaPaymentPage from "../../common/redirect/KlarnaPaymentPage.js";
-import PaymentResources from "../../data/PaymentResources.js";
-import {
-  goToShippingWithFullCart,
-  placeOrder,
-  proceedToPaymentAs,
-  verifyFailedPayment,
-  verifySuccessfulPayment,
-} from "../helpers/ScenarioHelper.js";
-import { PaymentDetailsPage } from "../pageObjects/plugin/PaymentDetails.page.js";
-
-const paymentResources = new PaymentResources();
-const user = paymentResources.guestUser.klarna.approved.nl;
-
-test.describe.parallel("Payment via Klarna Pay Later", () => {
-  test.beforeEach(async ({ page }) => {
-    await goToShippingWithFullCart(page);
-  });
-
-  test("should succeed", async ({ page }) => {
-    await proceedToPaymentAs(page, user);
-    const klarnaPaymentPage = await proceedToKlarnaPayLater(page);
-
-    await klarnaPaymentPage.makeKlarnaPayment("later", user.phoneNumber);
-    await verifySuccessfulPayment(page);
-  });
-});
-
-async function proceedToKlarnaPayLater(page) {
-  await new PaymentDetailsPage(page).selectKlarnaPayLater();
-  await placeOrder(page);
-  return new KlarnaPaymentPage(page);
-}
